Add ShareComponents.kt to share composable between AddNoteScreen and DetailNoteScreen Add Data Source,  Update Note done 90% (Image still have some bug)
