+++ app/build.gradle.kts
@@ -110,6 +110,7 @@ dependencies {
 
     implementation("com.google.accompanist:accompanist-placeholder-material:0.30.1")
 
+    implementation("androidx.compose.material:material-icons-extended:1.7.8") // Use latest Compose version
