+++ app/src/main/java/com/example/noteapp/data/data_source/local/database/NoteDatabase.kt
@@ -1,0 +1,12 @@
+package com.example.noteapp.data.data_source.local.database
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+
+@Database(entities = [NoteEntity :: class], version = 1, exportSchema = false)
+abstract  class NoteDatabase: RoomDatabase() {
+
+    abstract fun noteDao() : NoteDao
+
+}
