+++ app/src/main/java/com/example/noteapp/data/repository/ImageRepositoryImpl.kt
@@ -2,20 +2,26 @@ package com.example.noteapp.data.repository
 
 import android.content.Context
 import android.net.Uri
+import com.example.noteapp.data.data_source.local.source.ImageFileDataSource
 import com.example.noteapp.data.utils.ImageUtils
 import com.example.noteapp.domain.repository.IImageRepository
 import java.io.File
+import javax.inject.Inject
+import javax.inject.Singleton
 
-class ImageRepositoryImpl(private val context: Context) : IImageRepository {
+@Singleton
+class ImageRepositoryImpl @Inject constructor(
+   private val imageFileDataSource: ImageFileDataSource
+) : IImageRepository {
     override suspend fun saveImage(uri: Uri, fileName: String): String {
-        return ImageUtils.saveImageToInternalStorage(context = context, uri, fileName)
+        return imageFileDataSource.saveImage( uri, fileName)
     }
 
     override suspend fun getImage(fileName: String): File? {
-        return ImageUtils.readImageFile(context, fileName)
+        return imageFileDataSource.getImage(fileName)
     }
 
     override suspend fun deleteImage(fileName: String): Boolean {
-        return ImageUtils.deleteImageFile(context, fileName)
+        return imageFileDataSource.deleteImage(fileName)
     }
 }
