+++ app/src/main/java/com/example/noteapp/data/repository/NoteRepositoryImpl.kt
@@ -2,8 +2,8 @@ package com.example.noteapp.data.repository
 
 import android.util.Log
 import com.example.noteapp.common.Constants
-import com.example.noteapp.data.dao.NoteDao
-import com.example.noteapp.data.entity.NoteEntity
+import com.example.noteapp.data.data_source.local.database.NoteEntity
+import com.example.noteapp.data.data_source.local.source.NoteLocalDataSource
 import com.example.noteapp.domain.repository.INoteRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -12,13 +12,13 @@ import javax.inject.Singleton
 
 @Singleton
 class NoteRepositoryImpl @Inject constructor(
-    private val noteDao: NoteDao
+    private val  noteLocalDataSource: NoteLocalDataSource
 ): INoteRepository  {
     override suspend fun insertNote(note: NoteEntity): Long {
         Log.d(Constants.STATUS_TAG_ADD_NOTE_SCREEN,"Insert in NoteRepositoryImpl")
         return withContext(Dispatchers.IO){
             try {
-                noteDao.insertNote(note)
+                noteLocalDataSource.insertNote(note)
 
             }catch (e: Exception){
                 Log.d(Constants.ERROR_TAG_ADD_NOTE_SCREEN,"ERROR INSERT NOTE ${e.message}")
@@ -30,7 +30,7 @@ class NoteRepositoryImpl @Inject constructor(
     override suspend fun getAllNotes(): List<NoteEntity> {
         return withContext(Dispatchers.IO){
             try {
-                noteDao.getAllNotes()
+                noteLocalDataSource.getAllNotes()
 
             }catch (e: Exception){
                 Log.d(Constants.ERROR_TAG_ADD_NOTE_SCREEN,"ERROR GET ALL NOTE ${e.message}")
@@ -42,7 +42,7 @@ class NoteRepositoryImpl @Inject constructor(
     override suspend fun updateNote(note: NoteEntity): Int  {
         return withContext(Dispatchers.IO){
             try {
-                 noteDao.updateNote(note)
+                noteLocalDataSource.updateNote(note)
 
 
             }catch (e: Exception){
@@ -55,7 +55,7 @@ class NoteRepositoryImpl @Inject constructor(
     override suspend fun deleteNoteById(id: Long): Int {
         return withContext(Dispatchers.IO){
             try {
-                noteDao.deleteNoteById(id)
+                noteLocalDataSource.deleteNoteById(id)
 
 
             }catch (e: Exception){
@@ -68,7 +68,7 @@ class NoteRepositoryImpl @Inject constructor(
     override suspend fun getNoteByID(id: Long): NoteEntity? {
         return withContext(Dispatchers.IO){
             try{
-                noteDao.getNoteByID(id)
+                noteLocalDataSource.getNoteByID(id)
             }catch (e: Exception){
                 Log.d(Constants.ERROR_TAG_ADD_NOTE_SCREEN,"ERROR DELETE NOTE ${e.message}")
                 null
