+++ app/src/main/java/com/example/noteapp/data/repository/WorkManagerImpl.kt
@@ -1,0 +1,21 @@
+package com.example.noteapp.data.repository
+
+import android.annotation.SuppressLint
+import android.content.Context
+import androidx.work.WorkManager
+import com.example.noteapp.data.data_source.local.database.NoteEntity
+import com.example.noteapp.data.data_source.local.source.WorkManagerDataSource
+import com.example.noteapp.domain.repository.IWorkManager
+import javax.inject.Inject
+import javax.inject.Singleton
+
+
+@Singleton
+class WorkManagerImpl @Inject constructor(
+    private val workManagerDataSource: WorkManagerDataSource
+): IWorkManager {
+    @SuppressLint("NewApi")
+    override suspend fun scheduleNotification(context: Context, note: NoteEntity) {
+        workManagerDataSource.scheduleNotification(context, note)
+    }
+}
