+++ app/src/main/java/com/example/noteapp/di/RepositoryModule.kt
@@ -2,8 +2,10 @@ package com.example.noteapp.di
 
 import com.example.noteapp.data.repository.ImageRepositoryImpl
 import com.example.noteapp.data.repository.NoteRepositoryImpl
+import com.example.noteapp.data.repository.WorkManagerImpl
 import com.example.noteapp.domain.repository.IImageRepository
 import com.example.noteapp.domain.repository.INoteRepository
+import com.example.noteapp.domain.repository.IWorkManager
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -20,5 +22,10 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindImageRepository(imageRepositoryImpl: ImageRepositoryImpl): IImageRepository
+
+
+    @Binds
+    @Singleton
+    abstract fun bindWorkManagerRepository(workManagerImpl: WorkManagerImpl) :  IWorkManager
 }
