+++ app/src/main/java/com/example/noteapp/ui/presentation/detail/DetailViewModel.kt
@@ -1,6 +1,7 @@
 package com.example.noteapp.ui.presentation.detail
 
 import android.annotation.SuppressLint
+import android.content.Context
 import android.net.Uri
 import android.util.Log
 import androidx.compose.runtime.mutableStateOf
@@ -12,7 +13,9 @@ import com.example.noteapp.common.Constants
 import com.example.noteapp.domain.model.ItemDropMenu
 import com.example.noteapp.domain.model.Note
 import com.example.noteapp.domain.use_case.AddNoteUseCase
+import com.example.noteapp.domain.use_case.DeleteNoteUseCase
 import com.example.noteapp.domain.use_case.GetNoteUseCase
+import com.example.noteapp.domain.use_case.ScheduleNotifyUseCase
 import com.example.noteapp.domain.use_case.UpdateNoteUseCase
 import com.example.noteapp.ui.theme.high
 import com.example.noteapp.ui.theme.low
@@ -33,7 +36,10 @@ import javax.inject.Inject
 class DetailViewModel @Inject constructor(
     private val getNotesUseCase: GetNoteUseCase,
     private val updateNoteUseCase: UpdateNoteUseCase,
-    private val addNoteUseCase: AddNoteUseCase
+    private val addNoteUseCase: AddNoteUseCase,
+    private val scheduleNotifyUseCase: ScheduleNotifyUseCase,
+    private val deleteNoteUseCase: DeleteNoteUseCase
+
 ) : ViewModel() {
 
     val switchTopAppBar = mutableStateOf(false)
@@ -41,8 +47,6 @@ class DetailViewModel @Inject constructor(
     fun updateSwitchTopAppBar(updateValue: Boolean) {
         switchTopAppBar.value = updateValue
     }
-
-
 
 
     private val defaultNote = Note(0, "null", "null", "null", "null", 0, "null", "null", "null")
@@ -64,142 +68,89 @@ class DetailViewModel @Inject constructor(
         _note.value = updateValue
     }
 
-    // fun updateNote(note: NoteEntity): Int
-
-
 
     private val _detailState = MutableStateFlow<DetailState>(DetailState())
-    val detailState : StateFlow<DetailState> = _detailState.asStateFlow()
+    val detailState: StateFlow<DetailState> = _detailState.asStateFlow()
+
 
     @SuppressLint("NewApi")
-    fun updateNoteDatabase(note: Note){
-
+    fun updateNoteDatabase(note: Note, context: Context) {
         _detailState.value = DetailState(isLoading = true)
 
         viewModelScope.launch(Dispatchers.IO) {
-
             val currentTime = LocalDateTime.now()
             val today = LocalDate.now()
             val formatter = DateTimeFormatter.ofPattern("MMM dd, yyyy", Locale.ENGLISH)
-            val currentDate = today.format(formatter) // thời gian thêm note
-
+            val currentDate = today.format(formatter)
 
             val originalNote = getNotesUseCase.getNoteById(note.id)
-            Log.d(Constants.STATUS_TAG_DETAIL_SCREEN,"originalNote value = ${originalNote.toString()}")
+            var newImagePath: String? = note.image
 
-
-
-
-
-            if(selectedImageUri.value != null){
-                if(note.image!!.isNotEmpty()){
-
-                    val delete = updateNoteUseCase.deleteImage(note.image.toString())
-
-                    if( delete){
-
-                        val imagePath = if (selectedImageUri.value != null) {
-                            addNoteUseCase.saveImageToFileDir(
-                                selectedImageUri.value!!,
-                                currentTime.format(DateTimeFormatter.ofPattern("HH:mm:ss")) + ".Jpg"
-
-                            )
-
-                        } else {
-                            selectedImageUri.value.toString()
-                        }
-                        Log.d(Constants.STATUS_TAG_DETAIL_SCREEN,"Image name value = ${currentTime.format(DateTimeFormatter.ofPattern("HH:mm:ss")) + ".Jpg"}")
-
-
-                        val newNote = note.copy(
-                            image = imagePath,
-                            dateAdd = currentDate
-                        )
-
-
-                        if(newNote.title.isEmpty() || newNote.content.isEmpty()){
-                            _detailState.value = DetailState(isLoading = false, error = "ERROR: Fill in all fields. Please!")
-
-                        }else{
-                            updateNoteUseCase.updateNote(newNote)
-
-                        }
-
-                        _detailState.value = DetailState(isLoading = false, isSuccess = true )
-
-                        Log.d(Constants.STATUS_TAG_DETAIL_SCREEN,"Update Note have Image_Delete_Image, new note value = ${note.toString()}")
-
-                    } else{
-                        _detailState.value = DetailState(isLoading = false, error = "ERROR: Can't Update Note")
-
-
+            selectedImageUri.value?.let { uri ->
+                // Nếu có ảnh mới được chọn
+                if (!note.image.isNullOrEmpty()) {
+                    val deleted = updateNoteUseCase.deleteImage(note.image.toString())
+                    if (!deleted) {
+                        _detailState.value = DetailState(error = "ERROR: Can't Update Note")
+                        return@launch
                     }
                 }
-                else{
-                    val imagePath = if (selectedImageUri.value != null) {
-                        addNoteUseCase.saveImageToFileDir(
-                            selectedImageUri.value!!,
-                            currentTime.format(DateTimeFormatter.ofPattern("HH:mm:ss")) + ".Jpg")
 
-                    } else {
-                        selectedImageUri.value.toString()
-                    }
-                    Log.d(Constants.STATUS_TAG_DETAIL_SCREEN,"Image name value = ${currentTime.format(DateTimeFormatter.ofPattern("HH:mm:ss")) + ".Jpg"}")
-
-
-                    val newNote = note.copy(
-                        image = imagePath,
-                        dateAdd = currentDate
-                    )
-
-                    if(newNote.title.isEmpty() || newNote.content.isEmpty()){
-                        _detailState.value = DetailState(isLoading = false, error = "ERROR: Fill in all fields. Please!")
-
-                    }else{
-                        updateNoteUseCase.updateNote(newNote)
-
-                    }
-
-                    _detailState.value = DetailState(isLoading = false, isSuccess = true )
-
-                    Log.d(Constants.STATUS_TAG_DETAIL_SCREEN,"Update Note have Image, new note value = ${newNote.toString()}")
-                }
-            }else{
-                val newNote = note.copy(
-                    dateAdd = currentDate
+                newImagePath = addNoteUseCase.saveImageToFileDir(
+                    uri,
+                    currentTime.format(DateTimeFormatter.ofPattern("HH:mm:ss")) + ".Jpg"
                 )
-                if(originalNote!!.equals(newNote)){
+                Log.d(Constants.STATUS_TAG_DETAIL_SCREEN, "Saved new image: $newImagePath")
+            }
 
-                    _detailState.value = DetailState(isLoading = false, error = "ERROR: Nothing Change" )
+            val newNote = note.copy(
+                image = newImagePath,
+                dateAdd = currentDate
+            )
 
-                    Log.d(Constants.STATUS_TAG_DETAIL_SCREEN,"Khong co gia tri nao thay doi?")
-                }else{
-                    if(newNote.title.isEmpty() || newNote.content.isEmpty()){
-                        _detailState.value = DetailState(isLoading = false, error = "ERROR: Fill in all fields. Please!")
-
-                    }else{
-                        updateNoteUseCase.updateNote(newNote)
-                        _detailState.value = DetailState(isLoading = false, isSuccess = true )
-
-
-                    }
-
-                }
-                Log.d(Constants.STATUS_TAG_DETAIL_SCREEN,"Update Note don't have Image, new note value = ${newNote.toString()}")
-
+            if (newNote.title.isEmpty() || newNote.content.isEmpty()) {
+                _detailState.value =
+                    DetailState( error = "ERROR: Fill in all fields. Please!")
+                return@launch
             }
 
 
+            if (originalNote != null) {
+                if (
+                    selectedImageUri.value == null &&
+                    originalNote.title == newNote.title &&
+                    originalNote.content == newNote.content &&
+                    originalNote.category == newNote.category &&
+                    originalNote.priority == newNote.priority &&
+                    originalNote.image == newNote.image &&
+                    originalNote.dateNotify == newNote.dateNotify &&
+                    originalNote.timeNotify == newNote.timeNotify
+                ) {
+                    _detailState.value =
+                        DetailState(error = "ERROR: Nothing Change")
+                    return@launch
+                }
+            }
 
+            updateNoteUseCase.updateNote(newNote)
+            scheduleNotifyUseCase.scheduleNotification(context, note)
+            _detailState.value = DetailState(isLoading = false, isSuccess = true)
+
+            Log.d(Constants.STATUS_TAG_DETAIL_SCREEN, "Update completed: ${newNote}")
         }
-
     }
-
 
 
     //------------------------------------------------------------------///
 
-    private val defaultItemCategory = ItemDropMenu("Category", icon =  R.drawable.ic_category )
+    fun deleteNoteById(id: Long){
+        viewModelScope.launch(Dispatchers.IO) {
+           val delete = deleteNoteUseCase.deleteNoteById(id)
+
+        }
+    }
+
+    private val defaultItemCategory = ItemDropMenu("Category", icon = R.drawable.ic_category)
 
     val listCategory = listOf(
         ItemDropMenu("Work", icon = R.drawable.ic_work),
@@ -212,7 +163,6 @@ class DetailViewModel @Inject constructor(
     var categoryMenuExpanded = mutableStateOf(false)
 
     var selectedCategory = mutableStateOf(defaultItemCategory)//reset
-
 
 
     private val defaultItemPriority = ItemDropMenu("Priority", color = Color.Transparent)
@@ -229,22 +179,19 @@ class DetailViewModel @Inject constructor(
 
 
     var showPickerDate = mutableStateOf(false)
-    var selectedDate = mutableStateOf("00/00/0000")//reset
 
     var selectedImageUri = mutableStateOf<Uri?>(null) // reset
 
 
-
-
     var showDialog = mutableStateOf(false)
 
-    var dialogMessage =  mutableStateOf("") //reset
+    var dialogMessage = mutableStateOf("") //reset
 
-    fun updateShowDialog(updateValue: Boolean){
+    fun updateShowDialog(updateValue: Boolean) {
         showDialog.value = updateValue
     }
 
-    fun updateDialogMessage(updateValue: String){
+    fun updateDialogMessage(updateValue: String) {
         dialogMessage.value = updateValue
     }
 
@@ -253,7 +200,6 @@ class DetailViewModel @Inject constructor(
         categoryMenuExpanded.value = updateValue
 
     }
-
 
 
     fun updatePriorityMenuExpanded(updateValue: Boolean) {
@@ -269,15 +215,11 @@ class DetailViewModel @Inject constructor(
     }
 
 
-
-    fun updateSelectedDate(updateValue: String) {
-        selectedDate.value = updateValue
-    }
-
     fun updateSelectedImageUri(updateValue: Uri?) {
         selectedImageUri.value = updateValue
 
     }
+
